requirements.txt
kubernetes
python-digitalocean